chore(frontend): split helpers/toolbox.ts into a toolbox/ directory
Splits the growing helpers/toolbox.ts into cohesive modules under
helpers/toolbox/, with a barrel index.ts so the 25 existing
'@/helpers/toolbox' importers are unchanged:

- common.ts: constants (TOOLBOX_APP_KEY, TOOLBOX_ACTIONS) + general
  helpers (getGroupingActions, getStepGroupTypeAndCaption). Leaf module.
- for-each.ts: isForEachStep.
- if-then.ts: if-then predicates/branch extraction, useIfThenInitializer,
  getStepStructure, and the region-list helpers (StepRegion,
  buildRegionList, computeJumpTargets) — kept here because they depend on
  extractBranchesWithSteps.

Pure move: no behavioural change. The dependency direction is common <-
for-each <- if-then, so there are no import cycles.
